Split into separate modules and bring all public functions into `best`
